+++ packages/orq-rc/RELEASES.md
@@ -4129,3 +4129,13 @@ Based on:
 - [typescript v3.14.0-rc.12] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.12] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.12 - packages/orq-rc
+
+## 2025-10-07 06:57:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.635.1 (2.722.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.13] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.13] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.13 - packages/orq-rc